Change EveryMacInfo to Model
All information (for now) is going to come everymac.com, so I don't need multiple
sources. Additionally, the information gleaned from Everymac pertains to a specific
model of computer.

The short_model identifier has been removed from the Model object, as that value
is the responsibility of the SerialNumber: it comes directly from a computer's
serial number.